Change the updateCoin method in UserManger class and edit_profile_readandwrite method in WriteUser class.
Originally, the updateCoin method will directly store the amount of coin that receive to the user object. Now, this method will do calculation first and then store the coin to the user object.

the change of edit_profile_readandwrite method is as same as updateCoin method, it will do the calculation first and then change the amount of coin into file.