Pull down the WriteFile class from InterfaceAdapter layer as data access class to Framework & drivers layer.
Create a corresponding gateway class in Interface Adapter layer to deal with the request that send from user case and WriteFile class.

Create two new interfaces. One is GatewayInterface, which is in use case layer and implemented by gateway class. The Other one is DataAccessInterface, which is in InterfaceAdapter layer and implemented by three WriteFile classes.

Refactor all the varibale name and method name in all codes.